New message ConfigHash to sync device configs
When the device is brought online and has no configuration, the
device shall post a request to zedcloud with configHash set to
empty string (""). If the device has a valid configuration, then
the configHash would be set to the SHA256 checksum of the
configuration.

The zedcloud would consult it's deviceDB where it stores the
SHA256 checksum of the device configuration for all registered
devices, and figure if the device configuration for the device
has changed. If yes, then zedcloud would send back the device
configuration. If not, then zedcloud will respond with 304 HTTP
Status.