Addressing review feedback from PR#85
@rvs
Comment: why is this string("") and not just a string literal?
Resolution: Addressed in this PR

Comment:It would be useful to provide a comment with source of
        this manufacturer value to string mapping
Comment: Add a comment that these byte sequences are actually
         ASCII representations of vendor IDs.
Resolution: Added comment section above vendorRegistry

@eriknormark:
Comment: Where are you caching it?...
Resolution: Caching it in tpmmgr, under tpmHwInfo variable. For
            second Info message onwards, tpmmgr will not query
            TPM device, it will just return the cached value
            in tpmHwInfo. There is no code change for this in
            this PR, since it is already taken care in PR#85

Signed-off-by: Hariharasubramanian C S <cshari@zededa.com>