Currently, zedcloud copies phyIoDevice.Usage from devModel. Does not take into account that it can be over-ridden by device config.
SysAdapter.Uplink flag is set by reconciling Dev config and dev model.
So use this to sett the isMgmt flag.

THis allows us to change the usage of adapter in the interim - it was
how it was before model support came in.

The bigger changes to the API to remove overlapping fields will be done separately.

Signed-off-by: Kalyan Nidumolu <kalyan@zadeda.com>