Fixing downloading images progress bar
Get / GetAll always returned a shallow copy of the object from pubsub KeyMap.
But we had those Cast functions which converted into Json and back - thus creating a Deep Copy.
Somewhere in Jan - we got rid of this conversion and just started doing a static cast.
So, we are static casting the shallow copy and changing the copy. This caused the bug.

Signed-off-by: zed-rishabh <rgupta@zededa.com>