Fix `make installer-iso` error
Container `lfedge/eve-mkimage-iso-efi` was failing from `make
installer-iso` and reporting the error:

	standard_init_linux.go:211: exec user process caused "exec format error"

It was failing when entering the script `pkg/mkimage-iso-efi/make-efi`,
which is headed with `#!/bin/sh`.

Simply COPY'ing /bin/sh into the new container got past this error,
but encountered other errors when unable to found other utilities
used by the script `make-efi`.

Alpine's busyboxy package contains all of the depencies of `make-efi`,
so this commit simply adds that to the list of installed packages.

Busybox is about ~800KB.

An alternative approach would be to simply copy the individual binaries
and their deps, such as `cd`, `ln`, etc... used by make-efi.

Signed-off-by: Bobby Eshleman <bobbyeshleman@gmail.com>