Fix "0048-dell-xen-hack.patch" to make it compatible with mbox format
The patch "0048-dell-xen-hack.patch" is currently in raw form which works
perfectly with "patch" command but does not work with "git am". This change
fixes the patch so that it complies properly with mbox format so that git am
can consume it. All other patches are already in mbox format.

Signed-off-by: Ani Sinha <ani@anisinha.ca>