pillar: allocate qemu-overhead scaled to domain's memory
Currently we have a hard-coded amount of 600MiB, for any new VM no
matter what. This does not work for virtual machines with massive
amount of memory. Experimentally found that qemu requires about 2.4GiB
to run a Windows guest machine with 120GiB of RAM.

To address this situation we would allocate overhead for Qemu as 2.5%
of the total amount of RAM. But at least 600MiB to be on the safe side
and not to break any current scenarios.

Signed-off-by: Yuri Volchkov <yuri@zededa.com>