App CPU utilization needs to be scaled not by ticks but by vCPUs.
This ensures app instance so that the delta in CPU seconds will match
the delta in wallclock when the app keeps its vCPUs 100% busy.

To avoid loosing accuracy as we scale this PR is changing the local
field from seconds to nanoseconds. Separately we want to pass that
higher accuracy to the controller; it takes a while for a low
utilization app or EVE-OS to use one CPU second resulting in jagged
graphs.

Signed-off-by: eriknordmark <erik@zededa.com>
(cherry picked from commit dd81ac4e4e4f407a2e1866f99973a0f292309e99)