Make Xen device CPU reporting match KVM.
To be compatible with KVM, for Xen we need to add the guest CPU seconds
to the host CPU seconds. We also need to scale the resulting host CPU
seconds based on the number of CPUs, so that we report using at most 1
CPU second per second of wall clock time.

Signed-off-by: eriknordmark <erik@zededa.com>