Make robust against short power outages
To avoid reboot/crash/watchdog loops we do not use the saved config if
something bad caused a reboot. However, the determination of
BootReasonPowerFail does not seem to be robust against flaky power or
short power outages (I guess the disk/flash controller might not see a
full power cycle even though the OS reboots.) So allowing the use of the
saved config if BootReasonKernel to make offline mode more robust.

Signed-off-by: eriknordmark <erik@zededa.com>