Handle controller modifying the device UUID
Make sure all of the microservices which use the device UUID pick up the
change. Also detect the change when we get the config, or when we get a
404 from the /config API, and trigger the update. Should the device UUID
change we republish all info messages to the controller. Cleaned up the
misnamed RemoteTemporaryFailure variable and comments as part of
introducing the new SenderStatusNotFound.

Signed-off-by: eriknordmark <erik@zededa.com>
(cherry picked from commit a05da646af7cef5516ae2512db67cb18164adfe5)