Rename Airplane mode to Radio silence.
We have received feedback from EVE users that the term "Airplane-mode"
is not very fitting and possibly confusing. This is because in most likely
use-cases for this feature the edge node will not be actually inside an airplane
and the purpose of disabled radio transmission will not be related to or limited to
air travel safety.

In industrial applications, the act of disabling all radio transmission for safety
or security reasons is more commonly referred to as "Radio Silence"
(https://en.wikipedia.org/wiki/Radio_silence).

To change the name of the feature requires to modify already committed
APIs. Given that these APIs are fresh new and haven't been yet released,
this should not be a problem from the backward-compatibility point of view,
but we will have to ignore some yetus/buflint erros.

Signed-off-by: Milan Lenco <milan@zededa.com>
(cherry picked from commit 8b39c547fe5c0b04e06acc13e832c6a36b2c105e)