Fix confusing diag output
When the device is not yet onboarded diag (and ledmanager) can
incorrectly indicate that it is onboarded, which is very confusing. This
is due to the access to the old config API endpoint to get the UUID,
which can return StatusNotModified confusing EVE. Removing that API call
since it is no longer needed.

Signed-off-by: eriknordmark <erik@zededa.com>