Introduce IoBundle.KeepInHost
The semantics is that if KeepInHost is enabled, then device
remains in the host until it is assigned to an application (and goes
back to the host when it is unassigned).

We already have this behaviour for USB devices when debug.enable.usb
is enabled. Given that we need the same feature for wireless and vga devices,
it makes sense to generalize the implementation with a newly introduced
IoBundle.KeepInHost flag and use it for all these cases.

Also, the PR optimizes publishing of AssignableAdapters and ensures
that pubAssignableAdapters.Publish() is not called more often than it
needs to.

Signed-off-by: Milan Lenco <milan@zededa.com>