Generate device cert in tpmmgr
This simplifies the code and removes the dependency on openssh tools
(later we can move those tools to the debug container). As part of this
we also fix the CN in the generated cert, and backdate it by a day from
now for better rubustness against an inaccurate hardware clock.

Signed-off-by: eriknordmark <erik@zededa.com>