vhost: enable emulate_tpu to support scsi unmap
By enable option `emulate_tpu` in vhost iblock settings we activate
support scsi `unmap` command, which is an equivalent of trim/discard.

This command allows user to tell the block device that it does not
care about the context of a certain block, and hardware can adjust its
algorithms accordingly.

In our case this should tell ZFS that there is no need to
copy-on-write this block anymore and it can be used for something else.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>