ntpd container is not useful
If the clock is out of whack it does not slew it is in. This might be
due to it starting before the network interfaces is up and/or the -g
flag is not used by this container. In any case, this results in devices
which where the RTC is off by tens of minutes or more to not synchronize
their clock. Such an offset on the RTC appears to happen on some devices
which have been powered off for a long time, or devices which might have
had their RTC set to use some timezone other than UTC.

Signed-off-by: eriknordmark <erik@zededa.com>