[LVM] EVE Installer
new kernel cmd line variable eve_lvm_persist_size is introduced. It
can be specified in the same form as a size parameter for lvcreate
e.g. 30%VG

we use this variable to detect the fact that we are creating LVM volumes
and not ZFS pool. The magic ',' in eve_persist_disk is still required

RAID is not supported for LVM, instead all drives specified by
eve_persist_disk are added to one VG named eve_vg

Signed-off-by: Michael Malyshev <mikem@zededa.com>