With proxy certs the list of DPCs can easily exceed pubsub size limit
DevicePortConfig (DPC) may store one or more proxy certificates for each port.
With multiple DevicePortConfigs, kept in the DevicePortConfigList (DPCL) for
fallback purposes, this may quickly exceed the message size limit
of 64KB, as imposed by pubsub.
Normally, if the latest DPC is working (providing connectivity to the
controller), DPCL is being compressed before it is published.
The compression removes all obsolete (not latest) entries except
for lastresort. However, if the latest DPC is not working, DPCL
may just keep growing as new DPCs are being submitted from the controller.

A typicall scenario when this can be a problem, is when a user submits
invalid device port config (i.e. device will not use this failing latest config)
and then makes multiple unsucessful attempts to fix the config, each time
generating a new DPC. If the port configuration contains proxy certificates,
each with the size of 1-4 KBs, this can quickly adds up and exceed the pubsub
limit.

This PR tries to mitigate this issue.
First, using pubsub-large json tag, the pubsub system is ordered to send
certificates through temporary files and not through AF-UNIX socket
connections. DPCL message published via socket will contain references
to these files instead of certificate data.
For this the pubsub had to be enhanced to support pubsub-large tag
for fields which are under lists.
Second, an error informing about failed certification parsing, which
is stored into TestResults of a DPC, is trimmed to not contain an entire
certificate content.
Lastly, it makes sense to skip persisted DPCs with wrong certificates
and not carry them across device reboots.

Signed-off-by: Milan Lenco <milan@zededa.com>