Do not allow to impose radio silence during EVE update testing
If edge node is going through EVE update and radio silence is imposed
during the 10 minutes testing period, then the access to the controller
may be lost and device will fallback to the previous release.
This is in violation with the radio silence requirements, which state that
edge node should not trigger port config or EVE image fallback during a
(temporarily) imposed radio silence.
To prevent the EVE fallback from happening, zedagent will simply return error
back to the Local profile server if radio silence is requested during EVE
update testing period.

Signed-off-by: Milan Lenco <milan@zededa.com>