Reserve space to maintain performance in ZFS
It is recommended that storage usage should not go above 80% of
available space. Because, pool performance can degrade when a pool
is very full and file systems are updated frequently, such as on
a busy mail server. Full pools might cause a performance penalty.
If the primary workload is immutable files, then keep the pool in
the 95-96% utilization range. Even with mostly static content in the
95-96% range, write, read, and resolving performance might suffer.

This solution allows, allocate a dummy dataset with the size of 20%
of available space. This will take that space from the free pool, and
would not be used for other datasets/zvols. However, that space is
still available for copy-on-write.

Signed-off-by: Vitaliy Kuznetsov <ahil52_25@mail.ru>