kernel: make patches am-able
Only few of our patches can be applied using `git am`. This commit
turns all the patches into a proper format. This makes it a bit easy
to kick-start new kernel activity. One just needs to

   git am ~/eve/pkg/kernel/patches-5.10.x/*.patch

to have a clean git repo with all the patches in the history.

I tripple checked that old and new paches do have exactly the same
effect.

Signed-off-by: Yuri Volchkov <yuri@zededa.com>