(API only) Allow local profile server to purge/restart app instances
We have a customer request to make it possible to purge/restart running
application instances even when there is no connectivity between the
device and the controller. This is needed to solve the case of an app
being in a broken state and the user having no way of fixing it if the
device happens to lack external connectivity at that time.

Local purge/restart can be enabled by use of the Local Profile Server.
An existing endpoint api/v1/appinfo was extended to allow the server
to request commands to execute for locally running application
instances. Commands to run can be listed inside the message body
of the POST response.

This commit only defines the API of the feature. Implementation will
follow only once the API and the feature as such is approved.

Signed-off-by: Milan Lenco <milan@zededa.com>