zfs: performance patches
Some of these patches are already in the upstream ZFS repo, others are
in the process of getting there. We will remove at least some of these
patches with the next release of ZFS.

Patches significantly reduce the write amplification, therefore less
data needs to be written, and fewer uneccessary syncs issued.

Another feature introduced in the patches is write bandwidth smoothing
algorithm which prevents huge latency spikes under the heavy load on
the storage.

Signed-off-by: Yuri Volchkov <yuri@zededa.com>