embed ucode blobs into the kernel
We can see problems running VMs with intel microcode embedded (it is
clone of Ubuntu 18.04.6 from the bare-metal installation) because of no
possibility to update it from inside of VM.
If cpu microcode from inside of BIOS is outdated and have known
(possible) vulnerabilities, kernel of VM tries to update it if there are
 needed blobs inside /lib/firmware/intel-ucode (i.e. installed with
 intel-microcode package in Ubuntu) integrated into initramfs. And it
 stuck for our trial.
Removing of intel-microcode or adding dis_ucode_ldr to the kernel
options of VM helps to solve the problem. But in general we should use
the latest microcode to mitigate possible vulnerabilities.

In this commit I embed microcode blobs into the kernel

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>