introduce new mechanism of composing rootfs templates
This is preparation for introducing the development builds. The main
goal of the patch is to be able to combine different build flavors
together, by sequentially applying patches to the rootfs
Configuration.

At the moment the final rootfs configuration is built by applying
patches on the base configuration. To instruct the build to mutate
rootfs, the user needs to pass the `HV` variable to the `make`. For
example:

   make HV=acrn #to build the acrn-based Eve
   make HV=zfs-kvm #to build kvm-based with zfs as default file system

Currently, the build system would look under the `images` folder for
the corresponding patch. For example
`images/rootfs-acrn.yml.in.patch`. This means for each build flavor we
want to support we need a separate `*.patch` file.

This becomes a problem once we need to mix flavors. Particularly, with
upcoming “development” build flavor, which is orthogonal to the
hypervisor type and can be applied to kvm, xen and acrn. If we keep
this approach, to add a `dev` build we would have to introduce 3
patches under the `images` directory. This approach clearly does not
scale.

In this patch the `HV` string is tokenized, using minus (‘-’) as a
separator. And each token represents a separate patch under the
`images` directory. For example for the `HV=acrn-dev`, 2 patches would
be applied to the base configuration:

  - ‘images/acrn.patch’
  - ‘images/dev.patch’ (will be introduced later)

For now, since this new mechanism is not used, we have the luxury of a
transition period. This patch still keeps the old approach and the
both mechanisms co-exist: in addition to rootfs-*.yml.in also the
rootfs-*.yml.in.new file is generated and the context is
compared. Bail if they do not match.

If any broken build configurations are overlooked during testing,
keeping both mechanisms will let us catch any broken builds before
they cause any damage.

Signed-off-by: Yuri Volchkov <yuri@zededa.com>