wwan: Replace uqmi with qmicli
Using both uqmi and qmicli causes conflicts between connection
managementand location tracking. We are getting errors such as this:

[27 Apr 2022, 08:46:55] -Warning ** Cannot read from istream: connection broken

This is due to a limitation of the QMI protocol and the linux driver
qmi_wwan, which require that there is at most one client talking
to the modem. For this reason qmicli provides a proxy that multiplexes
multiple requests under one session. But uqmi is not able to use this
proxy, therefore we have to completely migrate from uqmi to qmicli.

Signed-off-by: Milan Lenco <milan@zededa.com>