Use kernel command line to deploy zfs in live img
We are forced to rebuild EVE image with zfs in version to use zfs with
live image. To generalize extraction of images from existing docker
image of EVE we should use options that we can configure with grub
options. This PR will deploy zfs persist partition when we will start
live image with eve_install_zfs_with_raid_level defined in kernel
cmdline.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>