Store vault key in TPM sealed and legacy
We can hit problems with storing of storage key because of no
connectivity to the controller or attestation template mismatch. In that
 case after reboot we will try to unseal key from TPM but if PCR values
 changed across reboots we will hit an error while unlock.

In this PR I store vault key in legacy mode (without sealing) alongside
with sealed one and use legacy key as a fallback. To properly handle
attestation process we should remove legacy key. I remove it on event
 of keys received from the controller, so we will wait for successful
 storing and responding of keys from the controller to not hit the
 problem with loosing of keys.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>