pubsub: populate from backup if the original json file is empty or missing
This is a follow up to [1].
This PR enhances pubsub such that it is now able to recover lost json files
with persisted items from backups. This is done by the Load()
method of the publisher's driver. The procedure of recovery is to first read
and then write the content of the *.bak file into the location of the original
file. This is done from inside of NewPublication() constructor, meaning that
all recovered items are then automatically also (re)published through sockets
(alongside all other loaded items).

[1] https://github.com/lf-edge/eve/pull/2630

Signed-off-by: Milan Lenco <milan@zededa.com>
(cherry picked from commit 71a6ed0790a8f786859c76172bed521c774f57ec)