Prefer MBIM protocol for SierraWireless EM7565
With cellular modem SierraWireless EM7565 we are experiencing quite
often a firmware bug causing the transmit queue to get stuck. In dmesg
we see:
  NETDEV WATCHDOG: wwan0 (qmi_wwan): transmit queue 0 timed out
Followed by a stacktrace which shows that netif_tx_lock hangs.
This can be only fixed by restarting the modem (not just the connection),
but it does not guarantee that it will not happen again and so we may end
up restarting the modem quite frequently, disrupting the ongoing traffic.
This bug is easily reproducible with the QMI control protocol.
With MBIM, we have not been able to reproduce this issue even after many
attempts. Therefore we prefer to use MBIM with this modem until we find
a better solution (that would not require running modem-specific AT
commands).

Signed-off-by: Milan Lenco <milan@zededa.com>