Update kernel to 5.10.121
We can reduce number of our patches with upstream support:
1. fget-Do-not-loop-with-rcu-lock-held removed because of no upstream
support and fixes inside https://github
.com/torvalds/linux/commit/054aa8d4 and
 https://github.com/torvalds/linux/commit/e386dfc5.
2. tpm-Rework-open-close-shutdown-to-avoid-races removed in favor of
https://github.com/torvalds/linux/commit/fb5abce6
3. rtl8821CU removed with update of library and also adds support for
arm64

It seems we may have endless loop inside the first patch in branch "if
(file->f_mode & mask)" without explicit setting of "again" to "false".
I can still see the problem handled by our patch for tpm in tests with
TPM-enabled Eden, so it is nice to have upstreamed version to fix this.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>