Introduce the delayed start feature.
An application configuration can include a delay setting. The setting contains
the amount of time (in seconds) EVE waits (after boot finish) before starting
the application.

Zedmanager is the first agent that handles the setting. It calculates the (wall)
time when the application should start. For the calculation, Zedmanager stores
the base time for counting the delays. The base time is defined as the moment
when Zedmanager first successfully get a configuration (either from a local
file or from a controller). Then Zedmanager calculates start time and stores it
in the application status.

Zedmanager, on any attempt to activate an application, first checks if the
current wall time is after the calculated start time stored in the status.
Suppose it's too early to start the applicated; zedmanager
change the application state to START_DELAYD and stops the handling.

Zedmanager, every 1 second, runs the check for the Applications in the
START_DELAYED state. If there is any, Zedmanager checks the current time, and
if it overcomes the calculated start time from the configuration, it triggers
the application modification handler. The handler tries to activate the
application again, and Zedmanager finally activates the application.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>