Make layout processing robust against partitions
We try to get disk from provided partition and operates on disks but it
blocks us from explicit setting of sda9 which we have in case of
installation on the same disk. We should explicitly provide partition,
for example /dev/sda9 to work with it for zfs state/layout processing
and avoid discrepancy.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>