Make /config read-only and without unique files
To enable measuring the config file system image at boot it must be read
only. In addition, we want identical hardware to have identical content
if possible. This means that files generated at install time like the
soft_serial file, the device.{cert,key}.pem, and tpm_credential should
not be in /config. When we have a TPM we get the certificate from TPM
NVRAM and ditto for the tpm_credential. The soft_serial and the cert+key
when there is no TPM are stored in /persist/. Note this means that if
you destroy and recreate /persist on a device without a TPM then the
device will generate a new identity.

Signed-off-by: eriknordmark <erik@zededa.com>