pillar: introduce debug build
Now if pillar is build with `DEV=y`, the debug symbols are not
stripped, and use reccomended GCFLAGS for devle.

Also include delve to the pillar container, and make required port
forwarded, so delve can attach.

Note that in this patch the development version of pillar container is
not picked up by linuxkit yet.

Signed-off-by: Yuri Volchkov <yuri@volch.org>