More sensible ticker range for location publishing to cloud
Currently, we use ticker with range (0.3 * interval, interval).
However, this means that the average interval is much less than
what is actually configured. This might be confusing for users.
Instead we therefore propose to use interval (0.5 * interval, 1.5 * interval),
averaging at the configured value.
This range was also shown by [1] to avoid the issue of self-synchronization.

[1] https://www.icir.org/floyd/papers/sync_94.pdf

Signed-off-by: Milan Lenco <milan@zededa.com>