use sparced write while rolling out image to zvol
It takes forever to write huge amount of data to the disk. The limit
for qemu-img operation was 16 minutes, which was not even close to be
enough for flashing 2TiB of data.

Assuming that zvol is created with all zerroes, we can instruct
qemu-img to skip writing zeroes. This drastically speeds up the
process of rolling out the image, as qcow2 almost never contains data
to fill target volume entirely.

Also in case we indeed have to flash a lot of data to the disk, this
patch increases timeout up to 2 hours.

Signed-off-by: Yuri Volchkov <yuri@volch.org>