Directories loops improvements
To not consume memory for the content of the directory we should read
several files and iterate over them, not over the whole directory.
This commit also includes improvements for newlogd to not store
unnecessary file path and to adjust limit for number of files inside
directories based on size limit and expected max size of file. Without
this limit we may go into enormous file count.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>