Fix diag output when proxy is configured
With explicit proxy configured, the resolution of the controller's
domain name is done by the proxy, not EVE. Actually, in some scenarious
it is possible that local DNS server(s) configured for uplink interfaces
are (intentionally) not able to translate external endpoints.
In such cases, it is expected for the DNS check performed by diag to
fail. Therefore it makes sense to skip this check if explicit proxy
is being used.

Signed-off-by: Milan Lenco <milan@zededa.com>