Fix handling of a disabled lastresort config
Even when lastresort is disabled by config, we might want to use it
in case there is no network configuration available. Recently [1],
a PR was merged, which enables lastresort config if persisted DPCL
is empty. However, with the first boot of a device, DPCL is obviously
empty, but a DPC can be submitted via config partition (override.json)
or through a specially formatted USB stick (usb.json). These DPCs are
submitted with some small delay, so they do not appear immediately
inside DPCL.
The idea of this PR is to change the current logic and let nim wait for
a minute for any DPC. If no config is available even after one minute
then it will forcefully enable lastresort.

[1] https://github.com/lf-edge/eve/pull/2620

Signed-off-by: Milan Lenco <milan@zededa.com>