Move per-device files back to /config
With these changes config will be read-only once the device has been
installed and the device certificate has been generated (the device cert
is generated either by the installer if the device has a real-time clock
or otherwise on first boot of EVE-OS). This assumes that the way will
will measure /config into a PCR will be able to skip the per-device
unique files to get the same PCR values across a fleet.

Signed-off-by: eriknordmark <erik@zededa.com>