Fix IP subnet obtained for wwan interface
When using netlink.AddrList(), we should check if the interface address
is reported with a Peer and if it is, use the subnet mask from the peer.
This is the case with Point-to-Point interfaces for example - the kind
of which EVE uses for wwan.

See: https://github.com/vishvananda/netlink/commit/b1cc70dea22210e3b9deca021a824f4edfd9dcf1

Signed-off-by: Milan Lenco <milan@zededa.com>
(cherry picked from commit c4a3aac3ad6c77748c297740ed1b5194af2dd5ac)