Recreate vault in case of no controller key
In case of no storage key stored on the controller side and no
information about previously sent storage keys, we are forced to
 re-create vault. We will send empty key to vaultmgr to indicate
 that no key received. If we receive empty key in vaultmgr and
 was not able to decrypt vault with sealed key, we remove old vault and
 setup the new one.

During installation, we create /persist/status/allow-vault-clean file
and remove it in case of success of publishing storage keys or if vault
unlocked and we about to create content tree or volume. In case of
non-unlocked vault, no storage keys received from controller and
existence of the file we will re-create the vault.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>