domainmgr: rename handle{Create,Modify} not related to pubsub
The name handleCreate and handleModify are used a lot in other parts
of pillar, epecially in pubsub. The pair touched in this patch has
nothing to do with pubsub, even though the naime strongly suggest it
does.

Rename those to something much more uniquie and less confusing:
doCreateDomain and doModifyDomain

Signed-off-by: Yuri Volchkov <yuri@volch.org>